feat: Restart slide-show when last/first item reached
'articleShown' value can not be more then # of the articles in array,
hen state of 'articleShown' reaches length of the array (-1),
it sets the articleShown state back to 0 & restarts the slide show.
Same principle applies to the 'slideLeft()' function,
with the difference that when the state 'articleShown' hits 0,
it goes back to the last article.